+++ roms_tools/tests/test_grid.py
@@ -1,10 +1,7 @@
 import pytest
 import numpy as np
 import numpy.testing as npt
-from scipy.ndimage import label
 from roms_tools import Grid
-from roms_tools.setup.topography import _compute_rfactor
-from roms_tools.setup.tides import TPXO
 import os
 import tempfile
 
@@ -109,73 +106,3 @@ class TestGridFromFile:
 
         finally:
             os.remove(filepath)
-
-
-class TestTopography:
-    def test_enclosed_regions(self):
-        """Test that there are only two connected regions, one dry and one wet."""
-
-        grid = Grid(
-            nx=100,
-            ny=100,
-            size_x=1800,
-            size_y=2400,
-            center_lon=30,
-            center_lat=61,
-            rot=20,
-        )
-
-        reg, nreg = label(grid.ds.mask_rho)
-        npt.assert_equal(nreg, 2)
-
-    def test_rmax_criterion(self):
-        grid = Grid(
-            nx=100,
-            ny=100,
-            size_x=1800,
-            size_y=2400,
-            center_lon=30,
-            center_lat=61,
-            rot=20,
-            smooth_factor=4,
-            rmax=0.2,
-        )
-        r_eta, r_xi = _compute_rfactor(grid.ds.h)
-        rmax0 = np.max([r_eta.max(), r_xi.max()])
-        npt.assert_array_less(rmax0, grid.rmax)
-
-    def test_hmin_criterion(self):
-        grid = Grid(
-            nx=100,
-            ny=100,
-            size_x=1800,
-            size_y=2400,
-            center_lon=30,
-            center_lat=61,
-            rot=20,
-            smooth_factor=2,
-            rmax=0.2,
-            hmin=5,
-        )
-
-        assert np.less_equal(grid.hmin, grid.ds.h.min())
-
-
-class TestTPXO:
-    def test_load_data_file_not_found(self):
-        # Test loading data from a non-existing file
-        with pytest.raises(FileNotFoundError):
-            TPXO.load_data("non_existing_file.nc")
-
-    def test_load_data_checksum_mismatch(self):
-        # Create a temporary file for testing
-        filename = "test_tidal_data.nc"
-        with open(filename, "wb") as file:
-            # Write some data to the file
-            file.write(b"test data")
-        # Test loading data with incorrect checksum
-        with open(filename, "wb") as file:
-            with pytest.raises(ValueError):
-                TPXO.load_data(filename)
-        # Remove temporary file
-        os.remove(filename)
